requirements.txt
requests~=2.31.0
python-qpid-proton~=0.39.0